Made window settings a protected variable in the game class
Removed game.enter as all it did was provide the main with references to something that was declared during runtime, instead i made a WindowConfig struct and made it so that it get set in the Game class constructor, I added Get Statements to abide by law of demeter.

This also reduces the obscurity of why width and height are declared in main and then set in game rather than game just simply owning it. It has one object that gets set and then has GetStatements for each variable

(CC G25, G36) (CCoreGuide C.9)