Added all render operations to the Renderer but all management to the RenderManager
I have made it so that the renderer is in charge of the actual rendering however the render manager handles in what order everything is supposed to happen. I have also bug fixed so that the movement is handled properly but needs to update and move on its own. Made so that the SDL_rects are created by the renderer before being displayed. Law of demeter is a pain in the ass